move command definition
