+++ cli/src/deposit.tsx
@@ -13,7 +13,8 @@ import { useCollateralBalance } from './pools/hooks/collateral.js'
 import { usePools, usePoolStats } from './pools/hooks/panoptic.js'
 import { type CollateralFullInfo } from './pools/hooks/common.js'
 import { NotificationContext } from './notification.js'
-import { CommandKeys, useCli, useOption, UserInputContext } from './commands.js'
+import { useCli, useOption, UserInputContext } from './commands.js'
+import { CommandKeys } from 'src/cmd.js'
 import { formatUnits, getContract } from 'viem'
 import { useERC20Balance } from './token.js'
 import { type AnnotatedTransaction, tryParseUnits } from './util.js'
