+++ cli/src/trade/trade.tsx
@@ -29,7 +29,8 @@ import { type PanopticPoolInfo } from '../pools/hooks/common.js'
 import { NotificationContext } from '../notification.js'
 import { formatUnits } from 'viem'
 import { type MarginUsage, useMarginEstimator } from './calc.js'
-import { CommandKeys, useCli, useOption } from 'src/commands.js'
+import { useCli, useOption } from 'src/commands.js'
+import { CommandKeys } from 'src/cmd.js'
 import CommandArgs from 'src/command-args.js'
 import { commandOptions } from 'src/options.js'
 import { usePools } from 'src/pools/hooks/panoptic.js'
